+++ app/src/main/java/com/dskja/betterstreamflix/support/SupportUrls.kt
@@ -7,6 +7,7 @@ package com.dskja.betterstreamflix.support
 object SupportUrls {
     const val BUY_ME_A_COFFEE_URL = "https://buymeacoffee.com/betterstreamflix"
     const val GITHUB_SPONSORS_URL = "https://github.com/sponsors/dskja"
+    const val PATREON_URL = "https://www.patreon.com/BetterStreamflix"
     const val TELEGRAM_URL = "https://t.me/BetterStreamflix"
     const val TELEGRAM_APP_URL = "tg://resolve?domain=BetterStreamflix"
     const val GITHUB_REPOSITORY_URL = "https://github.com/dskja/BetterStreamflix"